--- .vscode/extensions.json
@@ -1,10 +1,0 @@
-{
-  // See http://go.microsoft.com/fwlink/?LinkId=827846
-  // for the documentation about the extensions.json format
-  "recommendations": [
-    "editorconfig.editorconfig",
-    "mrmlnc.vscode-json5",
-    "rbbit.typescript-hero",
-    "syler.sass-indented",
-  ]
-}
