+++ common/constants.ts
@@ -1,0 +1,4 @@
+
+export const IPC = {
+  获取音乐列表: '获取音乐列表',
+}
