+++ electron/main/sql.ts
@@ -1,0 +1,23 @@
+
+export class Sql {
+
+  constructor() {
+
+  }
+
+  create() {
+
+  }
+
+  read() {
+
+  }
+
+  update() {
+
+  }
+
+  delete() {
+
+  }
+}
