+++ package.json
@@ -1,41 +1,35 @@
 {
-  "name": "electron-vite-react",
-  "productName": "Electron",
+  "name": "ply-music",
+  "productName": "PLY Music",
   "private": true,
   "version": "2.1.0",
   "description": "Vite React Electron boilerplate.",
-  "author": "草鞋没号 <308487730@qq.com>",
+  "author": "平亚兰 <308487730@qq.com>",
   "license": "MIT",
-  "main": "dist-electron/electron/main/index.js",
+  "main": "dist-electron/main/index.js",
   "scripts": {
     "dev": "vite",
     "build": "tsc && vite build && electron-builder"
   },
   "engines": {
-    "node": "^14.18.0 || >=16.0.0"
+    "node": "^16.0.0"
   },
-  "dependencies": {
-    "antd": "^5.1.7",
-    "react-router-dom": "^6.8.0",
-    "zustand": "^4.3.6"
-  },
+  "dependencies": {},
   "devDependencies": {
     "@types/react": "^18.0.21",
     "@types/react-dom": "^18.0.6",
-    "@vitejs/plugin-react": "^2.1.0",
-    "electron": "^21.1.0",
-    "electron-builder": "^23.3.3",
+    "@vitejs/plugin-react-swc": "^3.3.0",
+    "antd": "^5.4.2",
+    "electron": "^24.1.0",
+    "electron-builder": "^23.6.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.8.0",
     "sass": "^1.55.0",
-    "typescript": "^4.8.4",
-    "vite": "^3.2.2",
-    "vite-electron-plugin": "^0.5.0",
-    "vite-plugin-electron-renderer": "^0.10.2"
-  },
-  "debug": {
-    "env": {
-      "VITE_DEV_SERVER_URL": "http://127.0.0.1:7777"
-    }
+    "typescript": "^4.9.4",
+    "vite": "^4.2.1",
+    "vite-plugin-electron": "^0.11.1",
+    "vite-plugin-electron-renderer": "^0.13.14",
+    "zustand": "^4.3.7"
   }
 }
