--- src/assets/styles/app.module.scss
@@ -1,65 +1,0 @@
-.app {
-  text-align: center;
-
-  .appHeader {
-    background-color: #282c34;
-    min-height: 100vh;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    font-size: calc(10px + 2vmin);
-    color: white;
-
-    .logos {
-      display: flex;
-      box-sizing: border-box;
-      align-items: center;
-      padding: 0 5vw;
-      width: 100%;
-
-      .imgBox {
-        width: 33.33%;
-
-        .appLogo {
-          pointer-events: none;
-        }
-
-        @media (prefers-reduced-motion: no-preference) {
-          .appLogo {
-            animation: App-logo-spin infinite 20s linear;
-          }
-          @keyframes App-logo-spin {
-            from {
-              transform: rotate(0deg);
-            }
-            to {
-              transform: rotate(360deg);
-            }
-          }
-        }
-      }
-    }
-
-    button {
-      font-size: calc(10px + 2vmin);
-    }
-
-    .appLink {
-      color: #61dafb;
-    }
-
-    .staticPublic {
-      display: flex;
-      align-items: center;
-
-      code {
-        padding: 4px 7px;
-        margin: 0 4px;
-        border-radius: 4px;
-        background-color: rgb(30, 30, 30, .7);
-        font-size: 13px;
-      }
-    }
-  }
-}
