+++ src/main.tsx
@@ -1,13 +1,21 @@
-import React from 'react'
 import ReactDOM from 'react-dom/client'
-import App from './App'
-import './samples/node-api'
-import 'styles/index.css'
+import { ConfigProvider } from 'antd'
+import { RouterProvider } from 'react-router-dom'
+import router from './routes/router'
+import './ipc'
+
+import './main.css'
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>
+  <ConfigProvider
+    theme={{
+      token: {
+        colorPrimary: '#73d13d',
+      },
+    }}
+  >
+    <RouterProvider router={router} />
+  </ConfigProvider>
 )
 
 postMessage({ payload: 'removeLoading' }, '*')
